Move methods related to auto locking into a separate file
